# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EmailInput {
  email_body: string;
  subject: string;
  sender_email: string;
  /** @nullable */
  attachments?: string | null;
}

export interface ScoreBreakdownItem {
  category: string;
  score: number;
  max_score: number;
}

export interface SpamAnalysisDetails {
  suspicious_keywords: string[];
  email_structure_issues: string[];
  /** @nullable */
  sender_reputation_warning?: string | null;
  detected_attachments?: string[];
}

export interface SpamAnalysis {
  isSpam: boolean;
  confidence: number;
  spam_score: number;
  risk_factors: string[];
  score_breakdown: ScoreBreakdownItem[];
  details: SpamAnalysisDetails;
}

export interface ErrorResponse {
  error: string;
}