# lib/api-zod/src/generated/types/emailInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmailInput {
  email_body: string;
  subject: string;
  sender_email: string;
  /** @nullable */
  attachments?: string | null;
}