# lib/api-zod/src/generated/types/scoreBreakdownItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreBreakdownItem {
  category: string;
  score: number;
  max_score: number;
}