# lib/api-zod/src/generated/types/spamAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreBreakdownItem } from "./scoreBreakdownItem";
import type { SpamAnalysisDetails } from "./spamAnalysisDetails";

export interface SpamAnalysis {
  isSpam: boolean;
  confidence: number;
  spam_score: number;
  risk_factors: string[];
  score_breakdown: ScoreBreakdownItem[];
  details: SpamAnalysisDetails;
}