# lib/api-zod/src/generated/types/spamAnalysisDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpamAnalysisDetails {
  suspicious_keywords: string[];
  email_structure_issues: string[];
  /** @nullable */
  sender_reputation_warning?: string | null;
  detected_attachments?: string[];
}